Adds issues page and displays non-functioning issue form page
/issues loads a single issue from a list of issues for each user. It does not
fail gracefully, and it crashes if there is no entry in DB table users_issues.
It displays a link to the loaded issue, and future updates should display all.

/issue/:issue_id displays a single issue and three options, with another
non-option option (to ignore the issue). The options are presented as form
submit buttons, but are currently unprocessed. The template needs work to show
all appropriate DB columns on the view, along with processing tokens within
the dynamicly loaded options (i.e. random name).